The big leap
In 0.4.X there was a lot of parts to upgrade and the webapp internals
were flaky as hell. The goal of these changes is to rewamp data flow
inside the webapp and clean up outdated parts. Hopefully, this will
allow further development to be built on a more sane base.