Rename NOTIFICATION_EMAILS
Renamed NOTIFICATION_EMAILS to NOTIFICATION_EMAILS_THROTTLING

We have added a new setting for immediate email sending, so limit is no longer once per hour. Because of this, this setting is reduced to 0. But some admins may want to retain the limit so env variable is being retained.

Renaming to avoid accidental throttling when upgrading to new version as old name will exist in existing env file for existing servers.