Prevent notification email flood when enabled
If notifications are disabled, the last check is not set. This means if someone has notifications disabled then after years decide to enable them, they might get an email with years of notifications.

This commit means that when a user changes their notifications from none to another value, the last notification check time is reset so no prior notifications are sent.

If a user changes from, say, monthly to immediately, this logic is not triggered and they will receive an email with all the notifications they missed up to that point. This is intended.